Sam/requirements.txt
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.11.2
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
cmake==3.25.2
colorama==0.4.6
comtypes==1.1.14
cryptography==39.0.1
exceptiongroup==1.1.0
face-recognition==1.3.0
face-recognition-models==0.3.0
google==3.0.0
h11==0.14.0
idna==3.4
libretranslatepy==2.1.1
lxml==4.9.2
MouseInfo==0.1.3
numpy==1.24.2
opencv-python==4.7.0.68
outcome==1.2.0
pafy==0.5.5
Pillow==9.4.0
PyAudio==0.2.13
PyAutoGUI==0.9.53
pycparser==2.21
PyGetWindow==0.0.9
PyJWT==2.6.0
PyMsgBox==1.0.9
pyperclip==1.8.2
pypiwin32==223
PyRect==0.2.0
PyScreeze==0.1.28
PySocks==1.7.1
python-vlc==3.0.18121
pyttsx3==2.90
pytweening==1.0.4
pywin32==305
PyYAML==6.0
redis==4.5.1
requests==2.28.2
selenium==4.8.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SpeechRecognition==3.9.0
spotipy==2.22.1
translate==3.6.1
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.14
wikipedia==1.4.0
wincertstore==0.2
wsproto==1.2.0
